Reconfigure headers a little bit
